added alerts for successful function requests

+++ src/main/java/com/revature/service/RegistryServiceImpl.java
@@ -8,44 +8,61 @@ import com.revature.util.ConnectionFactory;
 public class RegistryServiceImpl implements RegistryService {
 
 	@Override
-	public void insertHero(String name, String id, String type) {
+	public boolean insertHero(String name, String id, String type) {
 		
-		String sql = "select hero_name from project1.superhuman ;";
-		ResultSet rs = ConnectionFactory.getResultSet(sql);
-		boolean exist = false;
-		// check to see if super human is already in registry
-		try {
-			while(rs.next()) {
-				if(rs.getString(1).equals(name))
-					exist = true;
-			}
-		} catch (SQLException e) {
-			e.printStackTrace();
-		}
+		boolean exist = findHero(name);
 		
 		if(!exist) {
 		System.out.println("inserting: " + name + " " + id + " " + type);
-		sql = "insert into project1.superhuman (hero_name, true_identity, alignmentid)"
+		String sql = "insert into project1.superhuman (hero_name, true_identity, alignmentid)"
 				+ " values ('" + name + "', '" + id + "', " + Integer.parseInt(type) + " )";
 	
 		ConnectionFactory.executeSQL(sql);
 		}
+		
+		return !exist;
 	}
 	
-	public void deleteHero(String name) {
+	public boolean deleteHero(String name) {
+		boolean success = findHero(name);
+		if(success) {
 		System.out.println("deleting: " + name);
 		String sql = "delete from project1.superhuman where hero_name = '" + name + "'";
 		
 		ConnectionFactory.executeSQL(sql);
+		}
+		
+		return success;
 	}
 	
 	@Override
-	public void UpdateHero(String name, String newName, String newid, String newType) {
+	public boolean UpdateHero(String name, String newName, String newid, String newType) {
+		boolean success = findHero(name);
+
+		if(success) {
+			String sql = "update project1.superhuman set hero_name = '" + newName + "', true_identity = '" + newid + 
+					"', alignmentid = " + Integer.parseInt(newType) + " where hero_name = '" + name + "'";
+			ConnectionFactory.executeSQL(sql);
+		}
 		
-		String sql = "update project1.superhuman set hero_name = '" + newName + "', true_identity = '" + newid + 
-				"', alignmentid = " + Integer.parseInt(newType) + " where hero_name = '" + name + "'";
-		ConnectionFactory.executeSQL(sql);
-		
+		return success;
+	}
+	
+	public boolean findHero(String name) {
+		String sql = "select hero_name from project1.superhuman ;";
+		ResultSet rs = ConnectionFactory.getResultSet(sql);
+		boolean success = false;
+		// check to see if super human is already in registry
+		try {
+			while(rs.next()) {
+				if(rs.getString(1).equals(name)) {
+					success = true;
+				}
+			}
+		} catch (SQLException e) {
+			e.printStackTrace();
+		}
+		return success;
 	}
